+++ covid_graphs/nginx/nginx.conf
@@ -60,7 +60,7 @@ http {
 
         location @proxy {
             include uwsgi_params;
-            uwsgi_pass dash_server:5000;
+            uwsgi_pass flask_server:5000;
         }
     }
 }
