+++ web/covid_web/rest.py
